test(e2e): move Go harness into dedicated folder
Keep the Go/Ginkgo e2e suite under tests/e2e/go so shell validators and manifests remain separate from Go harness packages.
