+++ tests/e2e/go/assertions/configmaps.go
@@ -12,7 +12,7 @@ import (
 	ginkgo "github.com/onsi/ginkgo/v2"
 	"github.com/onsi/gomega"
 
-	"github.com/NVIDIA/k8s-test-infra/tests/e2e/framework/kube"
+	"github.com/NVIDIA/k8s-test-infra/tests/e2e/go/framework/kube"
 )
 
 // ProfileConfigMaps ports the demo's "Profile ConfigMaps" step: when the fake
