+++ tests/e2e/go/assertions/nvidiasmi.go
@@ -13,9 +13,9 @@ import (
 	ginkgo "github.com/onsi/ginkgo/v2"
 	"github.com/onsi/gomega"
 
-	"github.com/NVIDIA/k8s-test-infra/tests/e2e/framework/kube"
-	"github.com/NVIDIA/k8s-test-infra/tests/e2e/framework/nodes"
-	"github.com/NVIDIA/k8s-test-infra/tests/e2e/profile"
+	"github.com/NVIDIA/k8s-test-infra/tests/e2e/go/framework/kube"
+	"github.com/NVIDIA/k8s-test-infra/tests/e2e/go/framework/nodes"
+	"github.com/NVIDIA/k8s-test-infra/tests/e2e/go/profile"
 )
 
 // NvidiaSMI ports validate-nvidia-smi.sh: nvidia-smi must run, and `-L` must
