+++ tests/e2e/go/assertions/nvlink.go
@@ -16,8 +16,8 @@ import (
 	ginkgo "github.com/onsi/ginkgo/v2"
 	"github.com/onsi/gomega"
 
-	"github.com/NVIDIA/k8s-test-infra/tests/e2e/framework/nodes"
-	"github.com/NVIDIA/k8s-test-infra/tests/e2e/profile"
+	"github.com/NVIDIA/k8s-test-infra/tests/e2e/go/framework/nodes"
+	"github.com/NVIDIA/k8s-test-infra/tests/e2e/go/profile"
 )
 
 var (
