+++ tests/e2e/go/assertions/pcisysfs.go
@@ -14,7 +14,7 @@ import (
 	ginkgo "github.com/onsi/ginkgo/v2"
 	"github.com/onsi/gomega"
 
-	"github.com/NVIDIA/k8s-test-infra/tests/e2e/framework/kube"
+	"github.com/NVIDIA/k8s-test-infra/tests/e2e/go/framework/kube"
 )
 
 // PCIDevicesDir is the fake /sys/bus/pci/devices tree render-pci-sysfs
