+++ tests/e2e/go/assets/assets.go
