+++ tests/e2e/go/assets/device-plugin-mock.yaml
