+++ tests/e2e/go/assets/gfd-mock.yaml
