+++ tests/e2e/go/assets/gpu-operator-values.yaml
