+++ tests/e2e/go/assets/kind-dra-config.yaml
