+++ tests/e2e/go/assets/kind-gpu-operator-config.yaml
