+++ tests/e2e/go/assets/kind-multi-node-config.yaml
