+++ tests/e2e/go/assets/validator-mock.yaml
