+++ tests/e2e/go/framework/kube/kube_test.go
